refine public api for IPC
