+++ Sources/SwiftAvroCore/IPC/Definitions.swift
