+++ Sources/SwiftAvroCore/IPC/Request.swift
@@ -23,23 +23,15 @@ import Foundation
 
 // MARK: - AvroIPCRequest
 
-/// Value-type, `Sendable` client-side IPC handler.
+/// Stateless, `Sendable` client-side IPC handler.
 ///
-/// Stateless with respect to sessions — all session state lives in the
-/// ``ClientSessionCache`` (`SessionCache<SessionRole.Client>`) actor passed to each method.
+/// Obtain via ``Avro/makeIPCRequest(clientHash:clientProtocol:context:)``
+/// rather than constructing directly.
 public struct AvroIPCRequest: Sendable {
 
     public let clientHash:     MD5Hash
     public let clientProtocol: String
 
-    /// Creates an `AvroIPCRequest`.
-    ///
-    /// - Parameters:
-    ///   - clientHash:     MD5 hash of the client's protocol.
-    ///   - clientProtocol: Protocol name string sent during retry handshake.
-    ///   - context:        Shared IPC context; when ``AvroIPCContext/knownProtocols``
-    ///     is non-nil the name is validated immediately and
-    ///     ``AvroHandshakeError/unknownProtocol(_:)`` is thrown if unrecognised.
     public init(
         clientHash:     MD5Hash,
         clientProtocol: String,
@@ -54,33 +46,30 @@ public struct AvroIPCRequest: Sendable {
 
     // MARK: - Handshake encoding
 
-    /// Encodes the initial handshake — `clientProtocol` is null on the first
+    /// Encodes the initial handshake — `clientProtocol` is nil on the first
     /// attempt per the Avro IPC specification.
-    /// Meta is always sent as an empty map rather than nil for structural consistency.
-    public func encodeInitialHandshake(context: AvroIPCContext) throws -> Data {
-        let avro = Avro()
+    public func encodeInitialHandshake(avro: Avro, context: AvroIPCContext) throws -> Data {
         avro.setSchema(schema: context.requestSchema)
         return try avro.encode(HandshakeRequest(
             clientHash:     clientHash,
             clientProtocol: nil,
             serverHash:     clientHash,
-            meta:           context.requestMeta  // always [:] rather than nil when empty
+            meta:           context.requestMeta
         ))
     }
 
     /// Encodes a retry handshake that includes the full client protocol string.
-    /// Meta is always sent as an empty map rather than nil for structural consistency.
     public func encodeHandshake(
+        avro:       Avro,
         serverHash: MD5Hash,
         context:    AvroIPCContext
     ) throws -> Data {
-        let avro = Avro()
         avro.setSchema(schema: context.requestSchema)
         return try avro.encode(HandshakeRequest(
             clientHash:     clientHash,
             clientProtocol: clientProtocol,
             serverHash:     serverHash,
-            meta:           context.requestMeta  // always [:] rather than nil when empty
+            meta:           context.requestMeta
         ))
     }
 
@@ -88,23 +77,22 @@ public struct AvroIPCRequest: Sendable {
 
     /// Decodes the server's handshake response and splits off the remaining payload bytes.
     public func decodeHandshakeResponse(
+        avro:    Avro,
         from data: Data,
-        context:   AvroIPCContext
+        context: AvroIPCContext
     ) throws -> (HandshakeResponse, Data) {
-        let avro = Avro()
-        let (response, next): (HandshakeResponse, Int) = try avro.decodeFromContinue(
-            from: data, schema: context.responseSchema
-        )
-        return (response, data.subdata(in: next ..< data.count))
+        let reader = avro.makeDataReader(data: data)
+        let response: HandshakeResponse = try reader.decode(schema: context.responseSchema)
+        return (response, data.suffix(reader.bytesRemaining))
     }
 
-    /// Resolves a handshake response against `cache`.
+    /// Resolves a handshake response, updating `context.clientCache` as needed.
     ///
     /// - Returns: Follow-up handshake data if the server responded with `.NONE`,
     ///   or `nil` if the handshake is complete (`.BOTH` / `.CLIENT`).
     public func resolveHandshakeResponse(
         _ response: HandshakeResponse,
-        cache:      ClientSessionCache,
+        avro:       Avro,
         context:    AvroIPCContext
     ) async throws -> Data? {
         switch response.match {
@@ -112,14 +100,14 @@ public struct AvroIPCRequest: Sendable {
             guard let serverHash = response.serverHash else {
                 throw AvroHandshakeError.noServerHash
             }
-            return try encodeHandshake(serverHash: serverHash, context: context)
+            return try encodeHandshake(avro: avro, serverHash: serverHash, context: context)
 
         case .CLIENT:
             guard let serverHash     = response.serverHash,
                   let serverProtocol = response.serverProtocol else {
                 throw AvroHandshakeError.noServerHash
             }
-            try await cache.add(hash: serverHash, protocolString: serverProtocol)
+            try await context.clientCache.add(hash: serverHash, protocolString: serverProtocol)
             return nil
 
         case .BOTH:
@@ -130,32 +118,26 @@ public struct AvroIPCRequest: Sendable {
     // MARK: - Call encoding
 
     /// Encodes a call request: metadata + message name + serialised parameters.
-    ///
-    /// Meta is always encoded (as empty map when none supplied) for structural consistency.
-    /// All parameter encoding is mandatory — throws immediately on any serialisation failure.
     public func encodeCall<T: Codable>(
+        avro:        Avro,
         messageName: String,
         parameters:  [T],
         serverHash:  MD5Hash,
-        cache:       ClientSessionCache,
         context:     AvroIPCContext
     ) async throws -> Data {
-        let avro = Avro()
         var data = Data()
-
-        // Metadata: always encode (empty map when none) — mandatory `try`
         data.append(try avro.encodeFrom(context.requestMeta, schema: context.metaSchema))
         data.append(try avro.encodeFrom(messageName, schema: AvroSchema(type: "string")))
 
-        // Empty message name = ping per Avro IPC spec: no schema lookup, no parameters.
+        // Empty message name = ping per Avro IPC spec: no parameters.
         guard !messageName.isEmpty else { return data }
 
-        // Schema validation: throw if schema is missing rather than silently skipping
-        guard let schemas = await cache.requestSchemas(hash: serverHash, messageName: messageName) else {
+        guard let schemas = await context.clientCache.requestSchemas(
+            hash: serverHash, messageName: messageName
+        ) else {
             throw AvroHandshakeError.missingSchema(messageName)
         }
         for (schema, parameter) in zip(schemas, parameters) {
-            // Mandatory encoding: `try` (not `try?`) so partial writes are impossible
             data.append(try avro.encodeFrom(parameter, schema: schema))
         }
         return data
@@ -165,97 +147,74 @@ public struct AvroIPCRequest: Sendable {
 
     /// Decodes a two-way response: metadata + error flag + response/error payload.
     public func decodeResponse<T: Codable>(
+        avro:        Avro,
         messageName: String,
         from data:   Data,
         serverHash:  MD5Hash,
-        cache:       ClientSessionCache,
         context:     AvroIPCContext
     ) async throws -> (ResponseHeader, [T]) {
-        let avro = Avro()
+        let reader = avro.makeDataReader(data: data)
 
-        let (hasMeta, metaEnd): (Int, Int) = try avro.decodeFromContinue(
-            from: data, schema: AvroSchema(type: "int")
-        )
+        let hasMeta: Int = try reader.decode(schema: AvroSchema(type: "int"))
         var meta: [String: [UInt8]]?
-        var flagOffset = metaEnd
         if hasMeta != 0 {
-            let (m, mEnd): ([String: [UInt8]]?, Int) = try avro.decodeFromContinue(
-                from: data.advanced(by: metaEnd), schema: context.metaSchema
-            )
-            meta       = m
-            flagOffset = metaEnd + mEnd
+            meta = try reader.decode(schema: context.metaSchema)
         }
 
-        let (flag, payloadOffset): (Bool, Int) = try avro.decodeFromContinue(
-            from: data.advanced(by: flagOffset), schema: AvroSchema(type: "boolean")
-        )
-        let absolutePayload = flagOffset + payloadOffset
+        let flag: Bool = try reader.decode(schema: AvroSchema(type: "boolean"))
         var params: [T] = []
 
         if flag {
-            if let errorSchemas = await cache.errorSchemas(hash: serverHash, messageName: messageName) {
-                var index = absolutePayload
+            if let errorSchemas = await context.clientCache.errorSchemas(
+                hash: serverHash, messageName: messageName
+            ) {
                 for (_, errorSchema) in errorSchemas {
-                    let (unionIndex, unionEnd): (Int, Int) = try avro.decodeFromContinue(
-                        from: data.advanced(by: index), schema: AvroSchema(type: "long")
-                    )
-                    index += unionEnd
+                    let unionIndex: Int = try reader.decode(schema: AvroSchema(type: "long"))
                     let schema = unionIndex > 0 ? errorSchema : AvroSchema(type: "string")
-                    let (param, paramEnd): (T, Int) = try avro.decodeFromContinue(
-                        from: data.advanced(by: index), schema: schema
-                    )
+                    let param: T = try reader.decode(schema: schema)
                     params.append(param)
-                    index += paramEnd
                 }
             }
         } else {
-            // Schema validation + mandatory throws: no try? — decode failure aborts cleanly
-            guard let responseSchema = await cache.responseSchema(
+            guard let responseSchema = await context.clientCache.responseSchema(
                 hash: serverHash, messageName: messageName
             ) else {
                 throw AvroHandshakeError.missingSchema(messageName)
             }
-            let param: T = try avro.decodeFrom(
-                from: data.advanced(by: absolutePayload), schema: responseSchema
-            )
+            let param: T = try reader.decode(schema: responseSchema)
             params.append(param)
         }
         return (ResponseHeader(meta: meta ?? [:], flag: flag), params)
     }
 }
-
+/*
 // MARK: - MessageRequest (backward-compatible façade)
 
-/// Manages the client side of the Avro IPC handshake and message encoding/decoding.
+/// Stateful client-side IPC façade.
 ///
-/// Thin adapter: stores the shared `context`, `cache`, and `request` value type,
-/// and forwards every call-site method to ``AvroIPCRequest``. Adds only what the
-/// value type cannot express: a raw `encodeHandshakeRequest` for hand-built
-/// `HandshakeRequest` values, and the dictionary-shaped `sessionCache` view.
+/// Obtain via ``Avro/makeIPCClient(clientHash:clientProtocol:context:)``.
+/// Holds a single `Avro` instance and forwards all calls to ``AvroIPCRequest``,
+/// passing `context` (which now owns the session cache).
 final class MessageRequest {
     let context:         AvroIPCContext
     private let avro:    Avro
     private let request: AvroIPCRequest
-    private let cache:   ClientSessionCache
     var clientRequest:   HandshakeRequest
 
-    /// Expose the underlying cache as the dictionary expected by tests.
     var sessionCache: [MD5Hash: AvroProtocol] {
-        get async { await cache.sessions }
+        get async { await context.clientCache.sessions }
     }
 
     init(context: AvroIPCContext, clientHash: MD5Hash, clientProtocol: String) throws {
         self.context = context
         self.avro    = Avro()
-        self.cache   = ClientSessionCache()
-        // AvroIPCRequest validates clientProtocol against context.knownProtocols.
         self.request = try AvroIPCRequest(
             clientHash: clientHash, clientProtocol: clientProtocol, context: context
         )
         self.clientRequest = HandshakeRequest(
-            clientHash: clientHash,
+            clientHash:     clientHash,
             clientProtocol: clientProtocol,
-            serverHash: clientHash,
+            serverHash:     clientHash,
             meta:           context.requestMeta
         )
         avro.setSchema(schema: context.requestSchema)
@@ -263,46 +222,44 @@ final class MessageRequest {
 
     // MARK: - Handshake
 
-    /// Raw encode for a hand-built `HandshakeRequest` — the one method the
-    /// value type does not provide because it owns its own request construction.
     func encodeHandshakeRequest(_ req: HandshakeRequest) throws -> Data {
         try avro.encode(req)
     }
 
     func initHandshakeRequest() throws -> Data {
-        try request.encodeInitialHandshake(context: context)
+        try request.encodeInitialHandshake(avro: avro, context: context)
     }
 
     func decodeResponse(from data: Data) throws -> (HandshakeResponse, Data) {
-        try request.decodeHandshakeResponse(from: data, context: context)
+        try request.decodeHandshakeResponse(avro: avro, from: data, context: context)
     }
 
     func resolveHandshakeResponse(_ response: HandshakeResponse) async throws -> Data? {
-        try await request.resolveHandshakeResponse(response, cache: cache, context: context)
+        try await request.resolveHandshakeResponse(response, avro: avro, context: context)
     }
 
     // MARK: - Session management
 
     func addSession(hash: MD5Hash, protocolString: String) async throws {
-        try await cache.add(hash: hash, protocolString: protocolString)
+        try await context.clientCache.add(hash: hash, protocolString: protocolString)
     }
 
     func removeSession(for hash: MD5Hash) async {
-        await cache.remove(for: hash)
+        await context.clientCache.remove(for: hash)
     }
 
     func clearSessions() async {
-        await cache.clear()
+        await context.clientCache.clear()
     }
 
     // MARK: - Call encode/decode
 
     func writeRequest<T: Codable>(messageName: String, parameters: [T]) async throws -> Data {
         try await request.encodeCall(
+            avro:        avro,
             messageName: messageName,
             parameters:  parameters,
             serverHash:  clientRequest.serverHash,
-            cache:       cache,
             context:     context
         )
     }
@@ -313,11 +270,12 @@ final class MessageRequest {
         from data:   Data
     ) async throws -> (ResponseHeader, [T]) {
         try await request.decodeResponse(
+            avro:        avro,
             messageName: messageName,
             from:        data,
             serverHash:  header.serverHash,
-            cache:       cache,
             context:     context
         )
     }
 }
+*/
